fix: mask every trace covered by copper pours
